+++ GaluhadPlugins/Pets/Strings/English/Pets.lua
@@ -387,35 +387,33 @@ _PETSTRINGS[310] = {[1]="Friendly Heron";[2]="Available in the Before the Shadow
 _PETSTRINGS[311] = {[1]="Mottled Craban";[2]="Obtained through crafting in Wildwood."}
 _PETSTRINGS[312] = {[1]="Golden Bat";[2]="Available by aiding Radagast at Rhosgobel."}
 
-_PETSTRINGS[500] = {[1]="Stripey Bat";[2]=""}
+_PETSTRINGS[313] = {[1]="Cozy Tabby Cat";[2]="Available to barter in Cardolan."}
+_PETSTRINGS[314] = {[1]="Red Salamander";[2]="Limited time availability: Legendary Item Reward Track Season 4"}
+_PETSTRINGS[315] = {[1]="Spooky Barrow-lurker";[2]="Available as a reward in Sarch Vorn, the Black Grave."}
 
+_PETSTRINGS[316] = {[1]="Black Marble Wolf-dog";[2]="Available from the Fall Festival."}
+_PETSTRINGS[317] = {[1]="White Goose";[2]="Available in the Yondershire.";}
+_PETSTRINGS[318] = {[1]="Warm Winter Goat Kid";[2]="Available from Season 4 of the Legendary Item Reward Track.";}
+_PETSTRINGS[319] = {[1]="Dusky Nether-hound";[2]="Available through barter by participating in Delvings.";}
+_PETSTRINGS[320] = {[1]="Brick Corgi";[2]="Available through barter by participating in Delvings.";}
+_PETSTRINGS[321] = {[1]="Friendly Stork";[2]="Available to barter in Cardolan.";}
+_PETSTRINGS[322] = {[1]="Friendly Egret";[2]="Limited time availability for Figments of Splendour.";}
+
+_PETSTRINGS[500] = {[1]="Stripey Bat";[2]=""}
 _PETSTRINGS[501] = {[1]="Injured Bat";[2]=""}
 _PETSTRINGS[502] = {[1]="Tan Bat";[2]=""}
-
 _PETSTRINGS[503] = {[1]="Black-wing Kite";[2]=""}
 _PETSTRINGS[504] = {[1]="Dusk-wing Kite";[2]=""}
 _PETSTRINGS[505] = {[1]="Maple Kite";[2]=""}
-
-_PETSTRINGS[506] = {[1]="Black Marble Wolf-dog";[2]="Available from the Fall Festival."}
-_PETSTRINGS[507] = {[1]="Tan Wolf-dog";[2]=""}
-
-_PETSTRINGS[508] = {[1]="White Goose";[2]="Available in the Yondershire."}
-
-_PETSTRINGS[509] = {[1]="Warm Winter Goat Kid";[2]="Available from Season 4 of the Legendary Item Reward Track."}
-_PETSTRINGS[510] = {[1]="Festive Yule Goat Kid";[2]=""}
-_PETSTRINGS[511] = {[1]="Ill Omened Beast";[2]=""}
-_PETSTRINGS[512] = {[1]="Ill Omened Crawler";[2]=""}
-
-_PETSTRINGS[513] = {[1]="Umber Nether-hound";[2]=""}
-_PETSTRINGS[514] = {[1]="Dusky Nether-hound";[2]="Available through barter by participating in Delvings."}
-
-_PETSTRINGS[515] = {[1]="Cream Corgi";[2]="";}
-_PETSTRINGS[516] = {[1]="Rust Corgi";[2]="";}
-_PETSTRINGS[517] = {[1]="Russet Corgi";[2]="";}
-_PETSTRINGS[518] = {[1]="Brick Corgi";[2]="Available through barter by participating in Delvings.";}
-_PETSTRINGS[519] = {[1]="Prepared Corgi";[2]="";}
-_PETSTRINGS[520] = {[1]="Library Corgi";[2]="";}
-_PETSTRINGS[521] = {[1]="Blonde Corgi";[2]="Available from Theodore Gorse during the Treasure Bugan event.";}
-_PETSTRINGS[522] = {[1]="Friendly Stork";[2]="Available to barter in Cardolan.";}
-_PETSTRINGS[523] = {[1]="Friendly Egret";[2]="Limited time availability for Figments of Splendour.";}
-_PETSTRINGS[524] = {[1]="Winter's Wind";[2]=""}
+_PETSTRINGS[506] = {[1]="Tan Wolf-dog";[2]=""}
+_PETSTRINGS[507] = {[1]="Festive Yule Goat Kid";[2]=""}
+_PETSTRINGS[508] = {[1]="Ill Omened Beast";[2]=""}
+_PETSTRINGS[509] = {[1]="Ill Omened Crawler";[2]=""}
+_PETSTRINGS[510] = {[1]="Umber Nether-hound";[2]=""}
+_PETSTRINGS[511] = {[1]="Cream Corgi";[2]=""}
+_PETSTRINGS[512] = {[1]="Rust Corgi";[2]=""}
+_PETSTRINGS[513] = {[1]="Russet Corgi";[2]=""}
+_PETSTRINGS[514] = {[1]="Prepared Corgi";[2]=""}
+_PETSTRINGS[515] = {[1]="Library Corgi";[2]="";}
+_PETSTRINGS[516] = {[1]="Blonde Corgi";[2]="Available from Theodore Gorse during the Treasure Bugan event.";}
+_PETSTRINGS[517] = {[1]="Winter's Wind";[2]="";}