+++ Changelog.md
@@ -1,4 +1,8 @@
 Change log:
+* 2022-11-16 - U34
+  * Added 12 pets, some old ones that were forgotten
+  * Mottled Craban, Golden Bat, Cozy Tabby Cat, Red Salamander, Spooky Barrow-lurker, Black Marble Wolf-dog, White Goose, Warm Winter Goat Kid, Dusky Nether-hound, Brick Corgi, Friendly Stork, Friendly Egret
+  * Future: Winter's Wind
 * 2022-09-22 - U33.2 - 4+2 new pets
   * Night Sky Carban, Lantern Gourd-lurker, Loyal Turtle, Friendly Heron
   * Future: Friendly Stork, Friendly Egret